Rerun code after sag changes size
The code that was last evaluated by the eval function (this happens by
clicking the “play” button) is rerun every time app.updateTwo is
triggered.

This way the drawing is redrawn in the “right” position.

Previously the user would have to click the play button every time the
window resized to make sure the drawing was in the right place, this is
history now.

**Note:** last evaluated code is stored in the variable
```app.lastEvaluatedCode```. This way we can easily access it any time
we need.

Also, I updated the update interval so the redrawing is done earlier.
(looks a bit smoother)