fixed menu-icon placement
The icon placement and size looked a bit weird when you toggled the
code section so that the icon-menu showed.

I made the menu-icon the same size as all the other icon-buttons and
made sure it was nicely aligned with the icon-menu.

Note: this issue was only on lg displays since all the other displays
don’t have the icon-menu feature.